+++ Source/PluginEditor.h
@@ -14,6 +14,7 @@
 #include "Data processing/mathGenerators.h"
 #include "Data processing/converters.h"
 #include "GUI/BasicLookAndFeel.h"
+#include "GUI/Simple elements/SimpleElements.h"
 
 
 struct SingleChannelFFTPathGenerator
@@ -90,67 +91,9 @@ private:
 
 
 
-
-struct RotarySliderWithLabels : juce::Slider
-{
-    RotarySliderWithLabels(juce::RangedAudioParameter& rap, const juce::String& unitSuffix) :
-    juce::Slider(juce::Slider::SliderStyle::RotaryHorizontalVerticalDrag,
-                 juce::Slider::TextEntryBoxPosition::NoTextBox),
-    param(&rap),
-    suffix(unitSuffix)
-    {
-        setLookAndFeel(&lnf);
-    }
-    
-    ~RotarySliderWithLabels()
-    {
-        setLookAndFeel(nullptr);
-    }
-    
-    struct LabelPos
-    {
-        float pos;
-        juce::String label;
-    };
-    
-    juce::Array<LabelPos> labels;
-    
-    void paint(juce::Graphics& g) override;
-    juce::Rectangle<int> getSliderBounds() const;
-    int getTextHeight() const { return 14; }
-    juce::String getDisplayString() const;
-private:
-    LookAndFeel lnf;
-    
-    juce::RangedAudioParameter* param;
-    juce::String suffix;
-};
 //==============================================================================
-struct PowerButton : juce::ToggleButton { };
 
-struct AnalyzerButton : juce::ToggleButton
-{
-    void resized() override
-    {
-        auto bounds = getLocalBounds();
-        auto insetRect = bounds.reduced(4);
-        
-        randomPath.clear();
-        
-        juce::Random r;
-        
-        randomPath.startNewSubPath(insetRect.getX(),
-                                   insetRect.getY() + insetRect.getHeight() * r.nextFloat());
-        
-        for( auto x = insetRect.getX() + 1; x < insetRect.getRight(); x += 2 )
-        {
-            randomPath.lineTo(x,
-                              insetRect.getY() + insetRect.getHeight() * r.nextFloat());
-        }
-    }
-    
-    juce::Path randomPath;
-};
+
 /**
 */
 class SimpleEQAudioProcessorEditor  : public juce::AudioProcessorEditor
